CenterNet2 train finished

+++ models/CenterNet2/inference_submission.py
@@ -31,22 +31,22 @@ except AssertionError:
 # config 불러오기
 cfg = get_cfg()
 add_centernet_config(cfg)
-cfg.merge_from_file('./configs/CenterNet2_R2-101-DCN-BiFPN_4x+4x_1560_ST.yaml')
+cfg.merge_from_file('output/CenterNet2/CenterNet2_R2-101-DCN-BiFPN_4x+4x_1560_ST(nms0.7test)/config.yaml')
 
 
 # config 수정하기
 cfg.DATASETS.TEST = ('coco_trash_test',)
-
-cfg.DATALOADER.NUM_WOREKRS = 2
-
-cfg.OUTPUT_DIR = './output'
-cfg.MODEL.WEIGHTS = os.path.join(cfg.OUTPUT_DIR, 'CenterNet2', 'CenterNet2_R2-101-DCN-BiFPN_4x+4x_1560_ST', 'model_final.pth')
-cfg.MODEL.RETINANET.SCORE_THRESH_TEST = cfg.MODEL.CENTERNET.INFERENCE_TH
-cfg.MODEL.ROI_HEADS.SCORE_THRESH_TEST = cfg.MODEL.CENTERNET.INFERENCE_TH
-if cfg.MODEL.META_ARCHITECTURE in ['ProposalNetwork', 'CenterNetDetector']:
-    cfg.MODEL.CENTERNET.INFERENCE_TH = cfg.MODEL.CENTERNET.INFERENCE_TH
-    cfg.MODEL.CENTERNET.NMS_TH = cfg.MODEL.ROI_HEADS.NMS_THRESH_TEST
-cfg.MODEL.PANOPTIC_FPN.COMBINE.INSTANCES_CONFIDENCE_THRESH = cfg.MODEL.CENTERNET.INFERENCE_TH
+#
+# cfg.DATALOADER.NUM_WOREKRS = 2
+#
+# cfg.OUTPUT_DIR = './output'
+# cfg.MODEL.WEIGHTS = os.path.join(cfg.OUTPUT_DIR, 'CenterNet2', 'CenterNet2_R2-101-DCN-BiFPN_4x+4x_1560_ST(nms0.7)', 'model_0016199.pth')
+# cfg.MODEL.RETINANET.SCORE_THRESH_TEST = cfg.MODEL.CENTERNET.INFERENCE_TH
+# cfg.MODEL.ROI_HEADS.SCORE_THRESH_TEST = cfg.MODEL.CENTERNET.INFERENCE_TH
+# if cfg.MODEL.META_ARCHITECTURE in ['ProposalNetwork', 'CenterNetDetector']:
+#     cfg.MODEL.CENTERNET.INFERENCE_TH = cfg.MODEL.CENTERNET.INFERENCE_TH
+#     cfg.MODEL.CENTERNET.NMS_TH = cfg.MODEL.ROI_HEADS.NMS_THRESH_TEST
+# cfg.MODEL.PANOPTIC_FPN.COMBINE.INSTANCES_CONFIDENCE_THRESH = cfg.MODEL.CENTERNET.INFERENCE_TH
 
 # model
 predictor = DefaultPredictor(cfg)
@@ -71,7 +71,7 @@ with open('../../dataset/test.json', 'r') as f:
 # output 뽑은 후 sumbmission 양식에 맞게 후처리
 prediction_strings = []
 file_names = []
-
+DEBUG = False
 class_num = 10
 color_dict = {i: tuple(map(int,np.random.choice(range(256), size=3))) for i in range(class_num)}
 cls_names = {i['id']:i['name'] for i in test_json['categories']}
@@ -90,18 +90,20 @@ for data in tqdm(test_loader):
     scores = outputs.scores.cpu().tolist()
 
     for target, box, score in zip(targets, boxes, scores):
-        prediction_string += (str(target) + ' ' + str(score) + ' ' + str(box[0]) + ' '
-                              + str(box[1]) + ' ' + str(box[2]) + ' ' + str(box[3]) + ' ')
+        if score > 0.1:
+            prediction_string += (str(target) + ' ' + str(score) + ' ' + str(box[0]) + ' '
+                                  + str(box[1]) + ' ' + str(box[2]) + ' ' + str(box[3]) + ' ')
+            if DEBUG:
+                img = cv2.rectangle(img, (int(box[0]), int(box[1])), (int(box[2]), int(box[3])), color_dict[target], 2)
+                cv2.putText(img, str(cls_names[target])+'  '+str(score)[:4], (int(box[0]), int(box[1]-10)), cv2.FONT_HERSHEY_SIMPLEX, 0.5, color_dict[target], 1)
 
-        img = cv2.rectangle(img, (int(box[0]), int(box[1])), (int(box[2]), int(box[3])), color_dict[target], 2)
-        cv2.putText(img, str(cls_names[target])+'  '+str(score)[:4], (int(box[0]), int(box[1]-10)), cv2.FONT_HERSHEY_SIMPLEX, 0.5, color_dict[target], 1)
-
-    cv2.imshow('', img)
-    if cv2.waitKey() == 27:
-        cv2.destroyAllWindows()
+    if DEBUG:
+        cv2.imshow('', img)
+        if cv2.waitKey() == 27:
+            cv2.destroyAllWindows()
 
     prediction_strings.append(prediction_string)
-    file_names.append(data['file_name'].replace('../dataset/', ''))
+    file_names.append(data['file_name'].replace('../../dataset/', ''))
 
 submission = pd.DataFrame()
 submission['PredictionString'] = prediction_strings
